Flash Messages
In the "Flash Messages" section, it seems to be a mistake.
The second portion of code (the one that shows after "read any flash messages from the session:") shows the legend ".. configuration-block::". I do not handle the notation of the document, but I'm pretty sure the problem is in the sintaxis.